Use the safe version of wcscpy
The reasoning behind the usage of wcscpy_s instead of the normal wcscpy is because the normal function will only just copy the wide string character into the source
without being aware of a possible consequence of a buffer overflow due to passing beyond the buffer boundaries. Since wcscpy_s counts the length number in
characters and not bytes, we'll use _countof().