implement acquire/release for ccflush
seems to fix various deadlock scenarios when writing in windows 11

what appears to happen is that ccflush magically acquires locks if we
don't acquire them explicitly. this then results in a deadlock where the
ccflush thread is holding the lock and waiting for writing to happen
(somewhere inside CcFlushCacheOneRange). meanwhile one of our threads is
stuck waiting for the lock (somewhere inside drv_write).